代码生成(非位域成员访问)

+++ src/symtab.rs
@@ -47,7 +47,7 @@ pub enum SymbolKind {
         storage_classes: Vec<StorageClass>,
     },
     Member {
-        bit_field: Option<BigInt>,
+        bit_field: Option<usize>,
     },
     Function {
         function_specs: Vec<FunctionSpec>,
